Fix migrate keys command
